moved things into components folder
