+++ frontend/src/components/Landing/About/About.jsx
