+++ frontend/src/components/Main/Submission/transcript.scss
