+++ frontend/src/components/SignUp/signUp.jsx
@@ -1,5 +1,5 @@
 import { Component } from "react";
-import '../SignUp/signUp.css';
+import './signUp.css';
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { Auth } from "aws-amplify";
